requirements.txt

sklearn

# for summary the network structure
torchinfo==1.7.0
# for use some metrics function
torchmetrics==0.9.2

# progress bar
tqdm==4.64.0

# some network strucutre API
pytorchvideo==0.1.5

# some hook used in pytorch lightning
lightning-bolts
